+++ src/main/java/edu/cnm/deepdive/join_me_backend/model/dao/InvitationRepository.java
@@ -5,7 +5,7 @@ import edu.cnm.deepdive.join_me_backend.model.entity.Square;
 import java.util.List;
 import org.springframework.data.repository.CrudRepository;
 
-public interface InvitationRepository extends CrudRepository<Invitation, Integer> {
+public interface InvitationRepository extends CrudRepository<Invitation, Long> {
 
   List<Invitation> findAll();
